Fit a matrix variate mixture model

Clustering by fitting a mixture model using EM with K groups and unconstrained covariance matrices for a matrix variate normal or matrix variate t distribution (with specified degrees of freedom nu).

matrixmixture(x, init = NULL, prior = NULL, K = length(prior),
  iter = 1000, model = "normal", method = NULL, tolerance = 0.1,
  nu = NULL, ..., verbose = 0, miniter = 3)

Arguments

x
data, p x q x n array

init
a list containing an array of K of p x q means, and optionally p x p and q x q positive definite variance matrices. By default, those are presumed to be identity if not provided. If init is missing, it will be provided using the prior or K by init_matrixmix.

prior
prior for the K classes, a vector that adds to unity

K
number of classes - provide either this or the prior. If this is provided, the prior will be of equal distribution.

iter
maximum number of iterations.

model
whether to use the normal or t distribution.

method
what method to use to fit the distribution. Currently no options.

tolerance
convergence criterion, using Aitken acceleration of the log-likelihood.

nu
degrees of freedom parameter

...
pass additional arguments to MLmatrixnorm or MLmatrixt

verbose
whether to print diagnostic output, by default 0. Higher numbers output more results.

miniter
minimum number of iterations

Value

A list of class MixMatrixModel containing the following components:

prior

the prior probabilities used.

init

the initialization used.

K

the number of groups

n

the number of observations

centers

the group means.

U

the between-row covariance matrices

V

the between-column covariance matrix

posterior

the posterior probabilities for each observation

pi

the final proportions

nu

The degrees of freedom parameter if the t distribution was used.

convergence

whether the model converged

logLik

a vector of the log-likelihoods of each iteration ending in the final log-likelihood of the model

method

the method used

call

The (matched) function call.

Examples

set.seed(20180221)
A <- rmatrixnorm(30,mean=matrix(0,nrow=3,ncol=4))
B <- rmatrixnorm(30,mean=matrix(2,nrow=3,ncol=4))
C <- array(c(A,B), dim=c(3,4,60))
prior <- c(.5,.5)
init = list(centers = array(c(rep(0,12),rep(2,12)), dim = c(3,4,2)),
             U = array(c(diag(3), diag(3)), dim = c(3,3,2)),
             V = array(c(diag(4), diag(4)), dim = c(4,4,2))
             )
res<-matrixmixture(C, init = init, prior = prior)
print(res) # note: prints head of posterior, not full list

#> $prior
#> [1] 0.5 0.5
#> 
#> $K
#> [1] 2
#> 
#> $n
#> [1] 60
#> 
#> $centers
#> , , 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.07351693 -0.08704738  0.13826607  0.06657800
#> [2,] -0.05562527 -0.08631881 -0.12719766 -0.07388405
#> [3,] -0.12802865 -0.33410177 -0.05220913 -0.03813562
#> 
#> , , 2
#> 
#>          [,1]     [,2]     [,3]     [,4]
#> [1,] 2.121336 2.230815 1.914284 2.082488
#> [2,] 2.328377 2.244675 1.569428 2.000457
#> [3,] 1.896128 1.800175 2.143966 2.048916
#> 
#> 
#> $U
#> , , 1
#> 
#>            [,1]       [,2]       [,3]
#> [1,] 1.00000000 0.14787429 0.02800157
#> [2,] 0.14787429 0.86714478 0.02258614
#> [3,] 0.02800157 0.02258614 0.98365621
#> 
#> , , 2
#> 
#>              [,1]        [,2]         [,3]
#> [1,]  1.000000000 -0.01739298 -0.004498719
#> [2,] -0.017392982  0.99877471 -0.100860756
#> [3,] -0.004498719 -0.10086076  0.977978554
#> 
#> 
#> $V
#> , , 1
#> 
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.81179286 -0.13501116  0.07550334 -0.15465746
#> [2,] -0.13501116  0.82941886 -0.04491670  0.06490662
#> [3,]  0.07550334 -0.04491670  1.03820594 -0.09435949
#> [4,] -0.15465746  0.06490662 -0.09435949  0.97277672
#> 
#> , , 2
#> 
#>             [,1]       [,2]       [,3]        [,4]
#> [1,]  0.76072996 0.01856626 0.03148176 -0.05903716
#> [2,]  0.01856626 0.67294240 0.05448756  0.10257452
#> [3,]  0.03148176 0.05448756 1.02451970  0.03848554
#> [4,] -0.05903716 0.10257452 0.03848554  1.03348310
#> 
#> 
#> $posterior
#>           [,1]         [,2]
#> [1,] 1.0000000 9.508947e-18
#> [2,] 0.9999999 1.072131e-07
#> [3,] 1.0000000 4.910925e-16
#> [4,] 1.0000000 7.500058e-18
#> [5,] 1.0000000 3.297751e-13
#> [6,] 1.0000000 1.630563e-08
#> 
#> $pi
#> [1] 0.5 0.5
#> 
#> $convergence
#> [1] TRUE
#> 
#> $iter
#> [1] 60
#> 
#> $logLik
#> [1] -4048.994 -3993.606 -3983.289 -3981.670 -3981.411
#> 
#> $model
#> [1] "normal"
#> 
#> $method
#> NULL
#> 
#> $call
#> matrixmixture(x = C, init = init, prior = prior)
#> 
#> attr(,"class")
#> [1] "MixMatrixModel"

if(interactive()) plot(res)
res<-matrixmixture(C, init = init, prior = prior, model = "t", nu = 5)
if(interactive()) plot(res)