Quadratic Discriminant Analysis for Matrix Variate Observations

See matrixlda: quadratic discriminant analysis for matrix variate observations.

matrixqda(x, grouping, prior, tol = 1e-04, method = "normal",
  nu = 10, ...)

Arguments

x
3-D array or list of matrix data.

grouping
vector

prior
a vector of prior probabilities of the same length as the number of classes

tol
by default, 1e-4. Tolerance parameter checks for 0 variance.

method
whether to use the normal distribution (normal) or the t-distribution (t). By default, normal.

nu
If using the t-distribution, the degrees of freedom parameter. By default, 10.

...
Arguments passed to or from other methods, such as additional parameters to pass to MLmatrixnorm (e.g., row.mean)

Value

Returns a list of class matrixqda containing the following components:

prior

the prior probabilities used.

counts

the counts of group membership

means

the group means.

U

the between-row covariance matrices

V

the between-column covariance matrices

lev

levels of the grouping factor

N

The number of observations used.

method

The method used.

nu

The degrees of freedom parameter if the t-distribution was used.

call

The (matched) function call.

Details

This uses MLmatrixnorm to find the means and variances.

See also

Examples

#' set.seed(20180221)
A <- rmatrixnorm(30,mean=matrix(0,nrow=3,ncol=4))
B <- rmatrixnorm(30,mean=matrix(1,nrow=3,ncol=4))
C <- array(c(A,B), dim=c(3,4,60))
groups <- c(rep(1,30),rep(2,30))
prior <- c(.5,.5)
D <- matrixqda(C, groups, prior)